requirements.txt
openai>=1.0.0
pyautogen>=0.2.0
langchain>=0.0.300
slack-bolt>=1.18.0
chromadb>=0.4.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pytest>=7.4.0